chore : comman entity 패키지 변경
